added a form that communicates with backend api
the form asks for first name, last name, password, email, credit card and cvv. The form isn't real so I made sure to tell people not to use their actual credit card information. The form then gets sent to the backend which sends it to the oracle database. (there are no password matches or attempts to validate the email yet, the frontend also doesn't display whether or not the request was successful).