added form options for postal code + address
The postal code only accepts G H J as the first character, it also does not accept any 0's. The length of the postal code must also be 6.

All of the errors are properly displayed so the user knows the format if the postal code is not properly formatted.

Also added an address section to the form, the address does not validate proper input. The only thing the address field will disallow is an empty string.